Split the delete, edit, and create editors modals into separate components
