+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import PageLayout from "./views/layout/PageLayout";
 import Login from './views/login/Login';
-import EditorsTable from './elements/editors/EditorsTable';
+import EditorsTable from './elements/editors/EditorsTable/EditorsTable';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import allReducers from './api/state/reducers';
