+++ src/components/core/core.scss
@@ -14,3 +14,13 @@ body {
 #root {
     height: 100%;
 }
+
+.modalButton {
+    margin-right: 1rem;
+}
+
+.errorMessage {
+    color: red;
+    margin-left: 3rem;
+    margin-bottom: 1rem;
+}
