+++ src/elements/editors/DeleteEditorModal/DeleteEditorModal.module.scss
@@ -1,0 +1,3 @@
+.deleteConfirmationMessage {
+    margin-bottom: 1rem;
+}
