+++ src/elements/editors/EditEditorModal/EditEditorModal.module.scss
@@ -1,0 +1,3 @@
+.formControlContainer {
+    margin-bottom: 1rem;
+}
