+++ src/elements/editors/EditorsTable/EditorsTable.module.scss
@@ -31,21 +31,3 @@ table {
     margin-left: 3rem;
     padding: 0.5rem;
 }
-
-.errorMessage {
-    color: red;
-    margin-left: 3rem;
-    margin-bottom: 1rem;
-}
-
-.deleteConfirmationMessage {
-    margin-bottom: 1rem;
-}
-
-.formControlContainer {
-  margin-bottom: 1rem;
-}
-
-.modalButton {
-    margin-right: 1rem;
-}
